Modals use Turbo for easier loading
Prior to this commit, the content of modals was pre-rendered and could be
shown either on load of the page or as requested from the page's script. This
commit changes the way this works to be more Turbo- and Stimulus-friendly.
Because Stimulus has lifecycle callbacks when targets are connected, we can
use those events to immediately pop up newly-loaded content. And because Turbo
can load new content into existing elements, we can use it to load this new
content.

The end result is that we now pop up a modal by loading its content on page
load, pre-rendered, and whether or not the page was already being displayed,
the modal will appear with the content in it.