Edit ProjectTeamMembership modal on Project Form
On page 2 of the Project form, we edit the team that is on the Project. The
existing members' memberships can be edited by clicking on the "edit" link.
This will pop up a modal that contains a form that allows the Institution,
role, and permissions to be edited.

This commit does not currently save the form, just displays it.