Allow Modals to close when removed
We add a hook to the ModalController so that, when a Modal is removed from the
document, it closes. This change also updates the close methods on the
controller so they can handle being called direct and being used as click
event handlers. This also updates the views to handle these `close` changes.