Use Turbo Streams for closing modal, showing edits
When using Turbo, the intent is to build a working setup with "traditional"
request/response full page loads first and then layer the dynamic updates from
Turbo on top. This commit uses the turbo-streams feature to update mutiple
elements at the same time from one request. When the form is submitted and the
save is successful, the stream renders "empty" into the modal (which, because
of a previous commit, closes the modal) and also renders an updated table into
the team members form.

Note: While this can be confusing, the updated table may contain the team
members in a different order, depending on their permissions. This is
intended.