Remove a project team member
This commit allows a team member to be removed from within the edit modal. The applicant can click a link to remove the team member and will need to confirm their decision before the destruction proceeds. If successful, the team member will be removed from the table of team members.