Create new user and membership from project form
This commit allows a new user to be created from the second page of the project form. A user can click a link and a modal will appear containing a form that can capture a subset of information required to create a new user and membership between the user and the current project. Due to the fact that minimal information is collected in this form, the remaining required information (address, terms acceptance, password, etc.,) is assigned using a fixed set of ‘dummy data’, which has been specifically requested by the client.

In addition to creating the user and team member, if the user creation fails, the transaction is rolled back and errors are displayed on the appropriate fields and the error is logged to the server. This work does not explicitly handle errors (other than logging) for failed membership as that is unlikely to happen if the creating the user is successful.

Note: This commit has a known CI failure that is addressed in the later commit titled "Specify modal dismissal better in System Specs".