Submit step 2 of new project form
This commit adds the markup to allow the button on the second page (Teams) of the project form to submit to the next step. It also adds the endpoint, , to route to. The markup on the Permits page is minimal (and without a presenter) but this is temporary and will be changing soon.