Fix the occasional failure generating a password
When the User is created on the User Form, we assign it a password using
`SecureRandom.urlsafe_base64(10)`, which makes a 10-character alphanumeric
string. But our requirements need a lowercase letter, uppercase letter,
number, and symbol. But a number wasn't always present in the `urlsafe_base64`
string. So this change tacks one on the end just like the `!` is.