Add and edit migrations to allow loading DB dump
This is not my favorite commit.

There is an issue where loading the DB dump and then running the migrations on
it causes a duplicate key error to happen even though there is not a unique
index, nor even duplicate data. The only way we could determine to fix it is
to remove the statement adding the fulltext index to `institutions`, which is
fine as we ended up not using it anyway.

To make this consistent, we also removed the other fulltext index definitions,
since we weren't using any of them. To safeguiard the migrations, we added a
new migration to remove the indices and accept that the statement might hold
an error if the index doesn't exist. We also changed the migrations that added
them to do nothing on up, but still remove them (if possible) on down. This
puts us in the best situation for not having these indices at all.

Now the migrations run from 0 to current and back without issue when the DB
dump is loaded (and also when it's not, of course).