Submit step 3 (permits) of project form
This commit modifies the relevant form objects and mark up in order to successfully submit the third page (the permits page) of the project form. The `ProjectPermitAnswersForm` (note the pluralization) responsible for generating and saving `ProjectPermitAnswerForm`s as well as the `approved_permits` attribute on the `project`. Due the architecture, there shouldn’t be a scenario in which any of these fields could fail a validation. This means that there is no real reason to display errors on this page of the form.