Add Funding model
This commit adds the `Funding` model, complete with the necessary validations and associations that are required to validate and eventually, submit a new `Funding` for a give `project`. It is worth calling out here that `reserve_id` is no longer needed in the `fundings` table, because of the new data model. However, in order to ensure that there are no issues with existing data, it is being left in. It is now allowable as null and the association is optional.