Refactor Form Error CSS Styles
We put the CSS into a mixin, and we also tweak the error generation in the
views so that instead of wrapping the label in a `div`, we put the error
messages into a `span` that is inside the `label` element. This allows the
error messages to more effectively share styles from the label in order to aid
alignment.