Present project funding in table
This commit allows funding that is added on the final page of the project form to be displayed in a table below the form that contains information about the funding title, agency, and award amount as well as a link to edit (but this doesn’t go anywhere yet). It also adds styles for the table.