Ensure funding award_amount is a number
This commit adds a validation to ensure the `numericality` of a `funding` instance's `award_amount`, allowing nil values. It also updates the `ProjectFundingForm` to add a setter that intercepts the incoming string value, strips it of non-digit characters (except decimals), converts it to an integer, and sets that value as the `fnuding.award_amount`. This ensures that by the time it reaches the model, it will pass the validation. This posed an issue previously because a value of `"1,000,000"` would be type cast and covered to `1.0`. It also did not account for the situation where someone would include the dollar sign which would yield `0.0` when cast.

This also comes after recognizing that the field in the database is represented as type `double(12,0)`, which is basically a floating point representation of an integer. We may want to change this field type in the future but for now, converting values to an integer and storing them with a `.0` suffix will suffice.