Allow submitting form when there are no Permits
So as it turns out, if there are no Permits available for the Project to
answer for, there is also no form for the user to submit. There is a message
telling them there are no Permits, but since there is no `<form>`, then button
that goes to the Funding page doesn't do anything (since there is nothing to
submit).

This commit puts the `<form>` into place so that even when nothing is asked,
there is a way for the process to continue.