Enable more Bullet reporting
This will let us know if we have a query-performance-related problem by
raising an error. Given that an uncaught error might tank the whole
application when uncaught on Heroku (because JawsDB limits based on the number
of queries per hour), we need to make sure this doesn't happen. Raising in
development when the error happens is a very obvious and unavoidable way for
us to be informed and thus fix the issue.