Fix the height of the Project form pages
In order to make sure the footer of the page is actually on the bottom of the
screen, this commit adjusts the flex and height of internal components to add
flexibility to the "main content" of the page. The footer will be attached to
the bottom of the window in all cases now.