Prevent missing params error in PermitAnswers
When there are no Permits that need to be addressed for a given project, the
form on the Permits page of the Project form will be empty. This is fine
except that the PermitAnswersController is expecting a `project` key to exist
in the params. We don't need it to contain anything, and, in fact, it doesn't
have to contain a "permitted" key at all. It just needs to exist.

This commit adds a "no_permits" key to the otherwise empty form so the
Controller doesn't freak out about its absence. There is a small test to
confirm that the POST request works.