Add & assign unknown institution to orphaned users
This commit adds a migration to create an institution that will be used for users that have no associated `institution_id`. In RAMS 3, there will not be a situation in which this can happen, but it was possible in RAMS 2 so we needed to address existing data. This data issue was causing the app to break when searching for users from the project form's team membership page because we were attempting to delegate `name` to non-existent `institution`s.