Handle autocomplete desc. for users with no email
This commit modifies the way that the autocomplete description for users is displayed when searching for a user by name on the project form's team membership page. There were two issues here that required attention. First, there was one user that represents a group of users that did not have an email address, which caused the method that masks an email to break. Now, when this result is returned, nil is returned so nothing is displayed. Additionally, this ensures that extra hyphens are not present by `compact`ing and `join`ing and array of strings rather than returning an interpolated string.