User search should search on all terms
When the User.search is called with a query like "kevin browne", then the
query that is constructed should take all terms in the string into account and
"AND" them together. Prior to this, "Kevin Browne" would search for everything
that contained "Kevin" and everything that contained "Browne". Now it searches
for things that contain "Kevin" AND "Browne".