Edit the project details
Users must be able to edit the project's details (especially after the project
has been created). This commit allows the user to get to the edit page if and
only if they have the `can_edit_project` permission. Even a PI that does not
have that permission will be unable to edit. If the attempt to load the edit
page fails, they are redirected to the project summary page.