Only allow authorized users to access project funding
This commit authorizes users before allowing them to view or manage funding for a project. In this case, a user should only be able to access the project funding page, or any of the endpoints that allow them to manage funding if they have a team_membership where active is true and can_edit_project is true.