Only allow authorized users to manage permits
This commit authorizes users before allowing them to view permits or create permit_answers for a project. In this case, a user should only be able to access the project permits page, or any of the endpoints that allow them to manage permits if they have a team_membership where `active` is true and `can_edit_project` is true. If they are not authorized, they will be redirected to the project summary page.

Note, this does not include the work to display reserve specific questions when editing, which will be included in a differet body of work.