Update reserve_questions table
This commit modifies the reserve_questions table (formerly ReserveQuestions) to use conventional syntax, and to add new columns as specified in this Slack [post](https://gnardog.slack.com/archives/C02BBA8BDN3/p1642714991001900) by @infomgr. This is necessary because these questions need to be displayed on the permits page when editing a project.