Allow ProjectPermitAnswers to be edited
In order to allow these to be edited in a way that doesn't require N+1 queries
to run, we modify the Form objects so that they use the MySQL-specific REPLACE
command. This command will attempt to INSERT the data, but if it runs afoul of
a key or unique index issue, it will DELETE the row and re-INSERT.

This means we can run one query that will ensure the answers are present in
all cases.