Include A11y matchers in Create/Edit project
There were a few changes needed to make the Create and Edit Project forms
properly accessible. Most were simply color changes to increase the contrast
ratio. Some wre making sure that there are alt attributes on all images. And
some are making sure that `id` uniqueness is being respected (this is
accomplished with the `namespace` argument to `form_with`.

Side note: The autocomplete plugin will not pass the `axe_clean` matcher
because it marks each opiton as `role="option"`, leading axe-matchers to
complain that the `ul` has things other than `li`s in it.