Update project status to open when completed
This commit serves to update a project's status from 'incomplete' to 'open' when the form to create a new project has been completed, i.e. the submit button on step 4 is clicked. Clicking the button, which is actually a form, will hit an endpoint that updates a `project`'s `status` to 'open' via a form object. If the update is successful, the user will be redirected to the project's summary (show) page. If not, the funding page will be rendered with a flash message stating that the `status` can not be updated.

The use of the `ProjectCompleteForm` may be overkill given that it is relatively simple. However, this will allow for greater flexibility in the future should other attributes need updating. It also follows the existing conventions for data management that we have established and allows for increased test coverage.

Lastly, the decision to construct and use `params` instead of passing in the `project` directly was not strictly necessary. Given that the data is being passed from controller to form object in order to modify a record, it seemed appropriate to use them and again allows for more flexibility in the future should other attributes need updating during this step.