Rename PermitsController to Questions/Answers
In order to have a more cohesively named form for answering both questions
raised by Permits and the questions raised by Reserves when applying for a
Visit, we are renaming the PermitsController to the QuestionsController. This
also renames the PermitAnswersController to AnswersController for similar
reasons.

This makes the URLs look like this:
/projects/:id/questions
/projects/:id/answers