Include preloading permits when showing Project
If the Project has multiple permit answers, then Bullet will detect that there
is an N+1 query when the statement column is loaded. To detect this in the
tests, we have to add another Permit to the process. And in order to fix the
N+1 query problem, we need to `includes` the permits when loading the answers.