Modify Answers query to be correct more often
When querying for ProjectPermitAnswers, the query that was in place would not
take the project being passed in into consideration. Upon attempting to narrow
the scope of results based on the project id and testing its correctness, we
found that permits that were answered for another project but not answered yet
for this project were not being returned. This forced us to rewrite the joins
that the query was using.

ActiveRecord does not allow for "extra" conditions on a left outer join, so we
had to write one manually via a raw SQL clause in the `joins` method. This
allowed us to specify the project's id in the join conditions rather than in a
WHERE clause. This means that if the join doesn't exist, NULL values will be
returned as required.