Adjust permit page styles
This commit tweaks the styles on the permits page of the new project form. Initially, this work was aimed at updating the questions to look more like an `<h3/>`. Those styles didn’t work _exactly_ with the design but I was able to make them more prominent, which was the goal. I also addressed some other inconsistencies	between the page and what was intended in the designs.