Always display something if answer is affirmative
This commit modifies the project summary page to ensure that there is always something displayed if a permit question was answered in the affirmative. Previously, if there was no statement, bullets for each answered question were being displayed with nothing beside them. Per the ticket:

> In the real world, this should not be an issue because the permit questions will just be copied into the statement field in the event that a permit is missing a value for that column.

This work basically ensures that the question will be shown as a fallback without requiring a data migration in the immediate.