Improve performance of User Autocomplete Queries
When typing in common names into the autocomplete search box ("Michael", etc),
the queries would often take a while to run, despite the fact that everthing
was indexed. As is turns out, querying on an unordered, flat list of ids in an
IN clause isn't terribly performant. Seeing as this is how Rails natively
handled the eager loading via `includes`, we wanted to find a better way.

Dealing with work we'd done previously, we knew that preloading data directly
(and manually) via specification in a `select` clause will aid performance. So
we added a small getter to the User class that will attempt to use a value
that we query for and then fallback to the "real" model if the cached balue
doesn't exist.

The end result is that instead of doing one query for Users and one
(potentially long) query for Institutions, we end up doing only one query for
Users (with a bit of Institution data included), which greatly aids loading
time assuming we stay within the expected use case.

And then as further reduction, we limit the query to 30 results.