Save ProjectPermit/ReserveAnswers in one query
In order to make sure we don't have an N+1 query when saving the Answers for
our Permit and Reserve questions, we are using the `INSERT ... ON DUPLICATE
KEY UPDATE` syntax for inserting the values into the table. This will allow us
to update the answer if the answer already exists and also insert new data if
it does not. It also updates the `updated_at` column instead of clobbering
the row as `REPLACE` does.