Presenters for Permits and ReserveQuestions
Because we are showing two types of question on the Questions page, we need a
presenter for each. This also improves the combined API so they are similar
and can be presented equally well in the views.

NOTE: In this commit, we add a method to the QuestionPresenter (and it also
exists in the PermitPresenter) that returns the Presenter when we ask for a
Form. This is because we're presenting this form differently than others. The
Presenters are loading extra data that doesn't normally exist on the models
it's loading. This data will not be on an equivalent Form object. It is also
due to the structure of the controllers, where the Questions controller
presents the questions and the Answers controller creates/updates the answers.
Creating a dedicated Form object for the Questions page -- or a dedicated
Presenter for the Answers would lead to more complication and muddy the
design. While using the Presenter as a Form also isn't the clearest, it's more
clear than the alternative.