Rename ProjectPermitAnswersForm and save Answers
This commit ties it all together. We rename the ProjectPermitAnswersForm to
ProjectAnswersForm because it handles both kinds of Answers, and then we
actually allow it to save both kinds of Answers. Despite the presence of an
answer_required column on ReserveQuestion, we don't actually check whether the
necessary value is present or not, because doing so would require us to make
more DB requests (plus everything should already have an answer).