Wait for elements to be in place when editing
The way that Capybara matchers work, they wait for the requested elements to
be in the DOM (or not be in the DOM with the `_no_` variants). The problem
with this is that Turbo is replacing the DOM with the same DOM in most cases,
so waiting for something to be (or not be) there can fail followup
expectations because the "old" element is the one that was found.

This is clearly suboptimal, but in order for Capybara to works as intended, we
need to make it explicitly wait. So, in this commit we add a small `sleep` so
Turbo can finish doing what it's doing before Capybara can check the DOM.