Fix wrapping institution search
This commit adjusts the styles of the autocomplete-results container such that it doesn't wrap - which causes the page to become very wide when there are a lot of results. These styles were previously defined in multiple places but now they live in a top level form mixin which can apply to all autocomplete select inputs used for searching.