Add Edit Amenity Rates for Manager Amenities and Rates page
This commit adds the ability to edit amenity reserves for the Amenities and Rates page. No styling yet.

Additionally, Bullet.raise does not play nice with `accepts_nested_attributes_for` and `has_many :amenity_rates, -> { in_order }`. And in the interest of making the deadline, it has been turned off, acknowledging that in the future I will work the warning messages.