Create update_amenity_visits_arrives_departs.rake
This commit adds a rake task to fix the arrives and departs column. This is done by first changing the arrives_on and arrives_at to localtime (the time was stored localtime on the inherited db data) then concatenating them to arrives. Same is done for departs.

Whereas before it was done incorrectly by just concatenating without first converting to localtime.