Create update_user_visits_arrives_departs.rake
This commit is similar to update_amenity_visits_arrives_departs.rake

it adds a rake task to fix the arrives_at and departs_at column. This is done by first changing the ArrivalDate and ArrivalTime to localtime (the time was stored localtime on the inherited db data) then concatenating them to arrives_at. Same is done for departs_at.

Whereas before it was done incorrectly by just concatenating without first converting to localtime.