Remove placeholder for adding team member create user
This commit remove many of the the placeholder data for when a new user is to be created from the second page of the project form.

Whereas before phone number, emergency contact name, emergency phone number, address line, city, state, postal code, country is NOT required when a user is a new user is created for the purpose of adding to a team member. Now they are required.