Add new fields for reserve personnel
This commit adds the field role_title and visible flag to the reserve personnel. With this commit, users are now able to attach a role title for reserve personnel such as Steward, Manager, Docent, etc. Users are now also able to hide reserve personnel from the public side of the site (for example, personnel such as accountant are not shown)