Fix reserve questions
This commit fixes multiple problems for reserve questions part of RAMS3.

First it fix the location of the answer. Whereas before both the visit and project answer are stored in one table, now it is stored in their respective answer table.

Second it fix the problem of past answered question/answer dissapearing if the setting of the question is updated in the present or future.