update reserve tag layout
This commit adds plus and minus icons to reserve category tag to indicate the shown or not shown reserve tags belonging to the category.

Furthermore, this commit remove javascript that is doing the showing and unshowing. Instead, we now use pure CSS to do this.

Lastly, the reserve category Ecosystem tags is now initially shown.