add outside reservation system
This commit adds an outside reservation system that can be set per individual amenity. When it is turned on for an amenity, any visit that uses the amenity will trigger a message and automatically forwards the user to a preset url. The url is set at the per reserve level.